% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot_styling.R
\name{forest_plot_styling}
\alias{forest_plot_styling}
\title{Styling for forest plots}
\usage{
forest_plot_styling(
  point.color = "#3333cc",
  pointrange.color = "#3333cc",
  refline.color = "gray40",
  refline.linetype = "dotted",
  refrange.color = NA,
  refrange.fill = "gray90"
)
}
\arguments{
\item{point.color}{Color for points}

\item{pointrange.color}{Color for point ranges}

\item{refline.color}{Color for reference line}

\item{refline.linetype}{Line type for reference line}

\item{refrange.color}{Color for border of reference range}

\item{refrange.fill}{Fill color for reference range}
}
\value{
An object of class \code{forest_plot_styling}.
}
\description{
This function provides the default styling for \code{\link[=forest_plot]{forest_plot()}}, which can be
overridden.
}
\examples{
forest_plot_styling()
}
\seealso{
\code{\link[=forest_plot]{forest_plot()}}.
}
