% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_forest_data.R
\docType{data}
\name{sample_forest_data}
\alias{sample_forest_data}
\title{Example dataset suitable for use with \code{\link[=forest_plot]{forest_plot()}}}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 32 rows and 11 columns.
}
\usage{
sample_forest_data
}
\description{
Example dataset suitable for use with \code{\link[=forest_plot]{forest_plot()}}
}
\keyword{datasets}
