% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{forest_plot}
\alias{forest_plot}
\title{Forest plot}
\usage{
forest_plot(
  fodat,
  scale = c("relative", "absolute"),
  lab = NULL,
  lim = NULL,
  refline = if (scale == "relative") 1 else NULL,
  refrange = if (scale == "relative") c(0.8, 1.25) else NULL,
  logscale = FALSE,
  widths = c(0.7, 0.3)
)
}
\arguments{
\item{fodat}{A \code{data.frame} in the format produced by \code{\link[=prepare_forest_data]{prepare_forest_data()}}.}

\item{scale}{Can be either "relative" or "absolute". Relative scale means
that values will be normalized by their reference value, while absolute
scale means values remain unchanged.}

\item{lab}{Label for the x-axis.}

\item{lim}{Limits for the x-axis.}

\item{refline}{Value of reference line.}

\item{refrange}{Limits for the reference range.}

\item{logscale}{Use log-scale for the x-axis?}

\item{widths}{A numeric vector of length 2 specifying the relative widths of
the main plot and the table containing the numeric values. By default, the
split is 70:30 for the main plot and the table.}
}
\value{
A \code{gtable} (see \code{\link[egg:ggarrange]{egg::ggarrange()}}).
}
\description{
Forest plot
}
\examples{
fodat <- sample_forest_data

# Use nice labels
covar.labels <- mappings::mapping(c(
        sex  = "Sex",
        wt   = "Body Weight (kg)",
        crcl = "Creatinine Clearance (mL/min)"))

param.labels = mappings::mapping(c(
        cmax = "C[max]",
        auc  = "AUC['0-8h']",
        cl   = "CL",
        vc   = "Vc"))

my_forest_plot <- function(.param, .covar=c("sex", "wt", "crcl")) {
  fodatplot <- data.table::setDT(fodat)[(param == .param) & (covar \%in\% .covar)]
  fodatplot[, covar := covar.labels(covar)]
  lab <- param.labels(unique(fodatplot$param))
  lab <- parse(text=paste("`Fold Change in`", lab, "`Relative to Reference`", sep="~"))
  forest_plot(fodatplot, scale="relative", lab=lab, logscale=TRUE)
}

my_forest_plot("cl")

my_forest_plot("vc", "wt")

my_forest_plot("cmax")

my_forest_plot("auc")

}
\seealso{
\code{\link[=prepare_forest_data]{prepare_forest_data()}}.
}
